requirements.txt
pygame>=2.5.0
azure-identity>=1.15.0
azure-cognitiveservices-speech>=1.35.0
